requirements.txt
flask
flask-sqlalchemy 
flask-login
flask-migrate
python-dotenv